refactor: extract dialogs and viewer
