+++ apps/web/src/App.tsx
@@ -1,136 +1,17 @@
 import { useState } from "react"
 import { Button } from "@/components/ui/button"
 import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog"
-import {
   SidebarProvider,
   Sidebar,
   SidebarContent,
   SidebarInset,
   SidebarRail,
 } from "@/components/ui/sidebar"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
 import logo from "@/assets/logo_dark.png"
-import { useResumeStore } from "@/hooks/use-resume-store"
+import { PdfViewer } from "@/components/pdf-viewer"
+import { PersonalInfoDialog } from "@/components/personal-info-dialog"
+import { JobInfoDialog } from "@/components/job-info-dialog"
 
-function PdfViewer() {
-  return (
-    <iframe
-      src="/blank.pdf"
-      className="h-full w-full border-0"
-      title="Resume Preview"
-    />
-  )
-}
-
-function PersonalInfoDialog({
-  open,
-  onOpenChange,
-}: {
-  open: boolean
-  onOpenChange: (open: boolean) => void
-}) {
-  const { userInfo, setUserInfo } = useResumeStore()
-  const [name, setName] = useState(userInfo.name ?? "")
-  const [email, setEmail] = useState(userInfo.email ?? "")
-
-  function handleSubmit(e: React.FormEvent) {
-    e.preventDefault()
-    setUserInfo({ name, email })
-    onOpenChange(false)
-  }
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Personal Information</DialogTitle>
-        </DialogHeader>
-        <form className="grid gap-4" onSubmit={handleSubmit}>
-          <div className="grid gap-2">
-            <Label htmlFor="name">Name</Label>
-            <Input
-              id="name"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              placeholder="Your name"
-            />
-          </div>
-          <div className="grid gap-2">
-            <Label htmlFor="email">Email</Label>
-            <Input
-              id="email"
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              placeholder="you@example.com"
-            />
-          </div>
-          <Button type="submit" className="mt-2 w-full">
-            Save
-          </Button>
-        </form>
-      </DialogContent>
-    </Dialog>
-  )
-}
-
-function JobInfoDialog({
-  open,
-  onOpenChange,
-}: {
-  open: boolean
-  onOpenChange: (open: boolean) => void
-}) {
-  const { jobInfo, setJobInfo } = useResumeStore()
-  const typedJobInfo = jobInfo as { title?: string; company?: string }
-  const [title, setTitle] = useState(typedJobInfo.title ?? "")
-  const [company, setCompany] = useState(typedJobInfo.company ?? "")
-
-  function handleSubmit(e: React.FormEvent) {
-    e.preventDefault()
-    setJobInfo({ title, company })
-    onOpenChange(false)
-  }
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Job Information</DialogTitle>
-        </DialogHeader>
-        <form className="grid gap-4" onSubmit={handleSubmit}>
-          <div className="grid gap-2">
-            <Label htmlFor="job">Job Title</Label>
-            <Input
-              id="job"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              placeholder="Desired role"
-            />
-          </div>
-          <div className="grid gap-2">
-            <Label htmlFor="company">Company</Label>
-            <Input
-              id="company"
-              value={company}
-              onChange={(e) => setCompany(e.target.value)}
-              placeholder="Company name"
-            />
-          </div>
-          <Button type="submit" className="mt-2 w-full">
-            Save
-          </Button>
-        </form>
-      </DialogContent>
-    </Dialog>
-  )
-}
 
 export default function App() {
   const [openPersonal, setOpenPersonal] = useState(false)
