+++ apps/web/src/components/pdf-viewer.tsx
@@ -1,0 +1,9 @@
+export function PdfViewer() {
+  return (
+    <iframe
+      src="/blank.pdf"
+      className="h-full w-full border-0"
+      title="Resume Preview"
+    />
+  )
+}
