+++ apps/web/src/components/personal-info-dialog.tsx
@@ -1,0 +1,58 @@
+import { useState } from "react"
+import { Button } from "@/components/ui/button"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog"
+import { Input } from "@/components/ui/input"
+import { Label } from "@/components/ui/label"
+import { useResumeStore } from "@/hooks/use-resume-store"
+
+export function PersonalInfoDialog({
+  open,
+  onOpenChange,
+}: {
+  open: boolean
+  onOpenChange: (open: boolean) => void
+}) {
+  const { userInfo, setUserInfo } = useResumeStore()
+  const [name, setName] = useState(userInfo.name ?? "")
+  const [email, setEmail] = useState(userInfo.email ?? "")
+
+  function handleSubmit(e: React.FormEvent) {
+    e.preventDefault()
+    setUserInfo({ name, email })
+    onOpenChange(false)
+  }
+
+  return (
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>Personal Information</DialogTitle>
+        </DialogHeader>
+        <form className="grid gap-4" onSubmit={handleSubmit}>
+          <div className="grid gap-2">
+            <Label htmlFor="name">Name</Label>
+            <Input
+              id="name"
+              value={name}
+              onChange={(e) => setName(e.target.value)}
+              placeholder="Your name"
+            />
+          </div>
+          <div className="grid gap-2">
+            <Label htmlFor="email">Email</Label>
+            <Input
+              id="email"
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              placeholder="you@example.com"
+            />
+          </div>
+          <Button type="submit" className="mt-2 w-full">
+            Save
+          </Button>
+        </form>
+      </DialogContent>
+    </Dialog>
+  )
+}
